pyproject.toml
[tool.poetry]
name = "ref-man-py"
version = "0.4.2"
description = "Ref Man Python Module"
authors = ["Akshay <akshay.badola.cs@gmail.com>"]
license = "GPL-3"
readme = "README.org"
homepage = "https://github.com/akshaybadola/ref-man-py"
classifiers=[
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    "Natural Language :: English",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Editors :: Emacs",
]
keywords = ['emacs', 'semanticscholar', 'bibliography', 'ref-man']

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
flask = "^1.1.2"
beautifulsoup4 = "^4.9.1"
psutil = "^5.8.0"
PyYAML = "^5.4.1"
lxml = "^4.6.4"
common_pyutil = "^0.8.0"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
poetry = "^1.1.8"
toml = "^0.10.2"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
test = "pytest:main"
ref_man = "ref_man_py.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
